Split mock_display.rs into a new module
